Make matching test data more likely for all_of quickcheck
The original condition was checking, for each iteration, if an arbitrary
f64 was greater than 0.1. Since f64 are 'signed', this would really have
a roughly 50% chance, which when multiplied by the number of expressions
would lead to very low likelihoods of generating matching test data.

Instead, a condition has been written that checks if an arbitrary u8 is
larger than 4. u8 has 256 values, so this represents a 1/64 chance of
failing the expression for every member.