Move combinator functions into expressions
In the process, most of the BoxedParser stuff has gone, and boxing has
moved to the Parser trait, along with an `as_ref` to offer a hopefully
more efficient option when a parser is being reused.

The export structure has been tweaked to accomodate everything. In
particular, the whole `expression` module is now public, and only the
functions and errors are re-exported at the top level.

Finally, links in the documentation have been changed to rely on the
nightly behaviour for 'intra rustdoc links'
(https://github.com/rust-lang/rfcs/blob/master/text/1946-intra-rustdoc-links.md#implied-shortcut-reference-links).