Remove the lifetime parameter from Parser
It turns out that what I really wanted here was [generic associated
types] e.g. an associated type with a lifetime parameter that can be
bound to the argument of a function. The desired interface was:

    trait Parser {
        type Value<'i>;
        fn parse<'i>(&self, &'i str) -> Self::Value<'i>;
    }

E.g., the `Value` associated type would be generic over a lifetime, and
parsers that wished to return a slice of the input could populate the
lifetime argument with the input's lifetime.

The naive approach of promoting the lifetime parameter to the trait is
called out as an [alternative] in the above RFC and is workable with the
caveat of the lifetime parameter 'infecting' all APIs that consume,
return and sometimes implement the trait. Given that the main
expressions that were using the lifetime – Any, Chr, and String – have
sensible alternatives, and that ultimately parsing a string and
returning itself is not very useful, removing the lifetime parameter
from the trait makes sense.

[generic associated types]: https://github.com/rust-lang/rfcs/blob/master/text/1598-generic_associated_types.md
[alternative]: https://github.com/rust-lang/rfcs/blob/master/text/1598-generic_associated_types.md#push-hrtbs-harder-without-associated-type-constructors