Implement `Parser` for `&str`
The dedicated `Chr` and `String` expressions have been replaced by an
implementation of Parser on `&str`. This makes most usages significantly
more concise by eliminating the need for a wrapping function. The one
narsty case is directly calling `Parser::parse` on a string. This
conflicts with `str::parse`, so needs to be spelled out:

    Parser::parse(&"...")

Note that the `&` is needed because `Parser` is implemented on `&str`,
and `Parser::parse` takes `&self`.